core/__init__.py


import hashlib
import os
import io
import re
import sys
import pytz
import uuid
import json
import time
import httpx
import errno
import discord
import ctypes
import winreg
import base64
import easygui
import logging
import discord
import webbrowser
import shutil
import random
import string
import asyncio
import zipfile
import datetime
import requests
import keyboard
import threading
import websocket
import importlib
import subprocess
import concurrent
import tls_client
import configparser
import concurrent.futures

session = tls_client.Session(
    client_identifier="chrome_116",
)

from .util.Plugins          import *
from .util.Headers          import *
from .util.RPC              import BackgroundRPC

from pystyle                import *
from ntpath                 import join
from time                   import sleep
from discord                import Embed
from multiprocessing        import Value
from genericpath            import isfile
from random                 import randint
from datetime               import datetime
from websocket              import WebSocket
from dataclasses            import dataclass
from twocaptcha             import TwoCaptcha
from colorama               import Fore, init
from json                   import dumps, loads
from distutils.version      import LooseVersion
from bs4                    import BeautifulSoup
from requests               import Session, session
from selenium               import webdriver, common
from concurrent.futures     import ThreadPoolExecutor
from urllib.request         import urlopen, urlretrieve
from discord.ui             import View, Modal, TextInput
from win32process           import GetWindowThreadProcessId
from win32gui               import GetWindowText, EnumWindows
from selenium.common        import exceptions as sel_exceptions
from psutil                 import Process, virtual_memory, cpu_count, disk_usage
from winreg                 import HKEY_LOCAL_MACHINE, OpenKey, CloseKey, QueryValueEx

# token
from .funcs.token.Formater            import Formatter
from .funcs.token.Checker             import Checker
from .funcs.token.Onliner             import Onliner
from .funcs.token.GuildJoiner         import TokenJoiner
from .funcs.token.GuildLeaver         import GuildLeaver
from .funcs.token.NameChanger         import NameChanger
from .funcs.token.PronChanger         import PronChanger
from .funcs.token.RemoveDuplicates    import RemoveDupes
from .funcs.token.BioChanger          import BioChanger
from .funcs.token.Nuker               import Nuker
from .funcs.token.Login               import Login
from .funcs.token.Info                import Info

# guild
from .funcs.server.ReplySpammer       import ReplySpammer
from .funcs.server.GroupSpammer       import GroupSpammer
from .funcs.server.ChannelSpammer     import ChannelSpammer
from .funcs.server.FriendReqSpammer   import RequestSpammer
from .funcs.server.GuildInfo          import GuildInfo

# other
from .funcs.other.Socials             import Socials

