requirements.txt
pytz
pypresence
httpx
discord
easygui
requests
keyboard
websocket-client  
tls-client  
pystyle
colorama
beautifulsoup4  
selenium